implement toggle temperature
